opps forgot to push this one, made a utils file for the interfaces
